+++ src/main/java/com/sprint/mission/discodeit/service/file/FileChannelService.java
@@ -8,11 +8,13 @@ import java.util.List;
 import java.util.UUID;
 
 public class FileChannelService implements ChannelService {
+
   private final ChannelRepository channelRepository;
 
   public FileChannelService(ChannelRepository channelRepository) {
     this.channelRepository = channelRepository;
   }
+
   @Override
   public Channel create(String name) {
     Channel channel = new Channel(name);
@@ -31,16 +33,21 @@ public class FileChannelService implements ChannelService {
   }
 
   @Override
-  public void update(UUID id, String newName) {
+  public Channel update(UUID id, String newName) {
     Channel channel = channelRepository.findById(id);
     if (channel != null) {
       channel.updateName(newName);
       channelRepository.save(channel);
     }
+    return channel;
   }
 
   @Override
-  public void delete(UUID id) {
-    channelRepository.delete(id);
+  public Channel delete(UUID id) {
+    Channel channel = channelRepository.findById(id);
+    if (channel != null) {
+      channelRepository.delete(id);
+    }
+    return channel;
   }
 }
